Make some code changes as a precursor for unit-testing
